update 9 weather condition for each scenarios
